+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/ExcludeKitTest.groovy
@@ -124,7 +124,7 @@ class ExcludeKitTest extends AbstractKitTest {
         build("""
             plugins {
                 id 'groovy'
-                id 'ru.vyarus.animalsniffer' version '1.4.2'
+                id 'ru.vyarus.animalsniffer' version '1.4.6'
                 id 'ru.vyarus.quality'
             }