+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/ExcludeSourcesKitTest.groovy
@@ -114,7 +114,7 @@ class ExcludeSourcesKitTest extends AbstractKitTest {
         build("""
             plugins {
                 id 'java'
-                id 'ru.vyarus.animalsniffer' version '1.4.2'
+                id 'ru.vyarus.animalsniffer' version '1.4.6'
                 id 'ru.vyarus.quality'
             }