+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Gradle 4.10 compatibility:
+    - updated com.github.spotbugs plugin 1.6.2 -> 1.6.3  
+
 ### 3.1.1 (2018-07-22)
 * Fix maven central and jcenter artifacts (missed dependency) (#9)
 * Update spotbugs 3.1.5 -> 3.1.6